Create Quaternions.jl
remove quaternions from generic tests
